moved add timeline event handler to its own func
